# contestr/api/openapi.yaml
openapi: 3.0.0
info:
  title: API контестов
  description: Ручки для получения данных о контестах
  version: 1.0.0
paths:
  /api/contests:
    get:
      operationId: getContests
      responses:
        200:
          description: Список зарегистрированных контестов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RegisteredContestItem'

  /api/admin/contests:
    get:
      operationId: getAdminContests
      security:
        - bearerAuth: []
      responses:
        200:
          description: Список зарегистрированных контестов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RegisteredContestItem'
        401:
          $ref: '#/components/responses/ErrorResponse'
    post:
      operationId: postAdminContest
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContestRequest'
      responses:
        201:
          description: Контест зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisteredContestItem'
        400:
          $ref: '#/components/responses/ErrorResponse'
        401:
          $ref: '#/components/responses/ErrorResponse'
        409:
          $ref: '#/components/responses/ErrorResponse'
        502:
          $ref: '#/components/responses/ErrorResponse'

  /api/admin/contests/{contest_id}:
    delete:
      operationId: deleteAdminContest
      security:
        - bearerAuth: []
      parameters:
        - name: contest_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        204:
          description: Контест удалён из реестра
        401:
          $ref: '#/components/responses/ErrorResponse'
        404:
          $ref: '#/components/responses/ErrorResponse'

  /api/admin/contests/{contest_id}/handles:
    get:
      operationId: getAdminContestHandles
      security:
        - bearerAuth: []
      parameters:
        - name: contest_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Маппинги handle → имя
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CodeforcesHandleItem'
        401:
          $ref: '#/components/responses/ErrorResponse'
        404:
          $ref: '#/components/responses/ErrorResponse'
    put:
      operationId: putAdminContestHandles
      security:
        - bearerAuth: []
      parameters:
        - name: contest_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutCodeforcesHandlesRequest'
      responses:
        200:
          description: Сохранённые маппинги
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CodeforcesHandleItem'
        400:
          $ref: '#/components/responses/ErrorResponse'
        401:
          $ref: '#/components/responses/ErrorResponse'
        404:
          $ref: '#/components/responses/ErrorResponse'

  /api/admin/contests/{contest_id}/handles/{handle}:
    delete:
      operationId: deleteAdminContestHandle
      security:
        - bearerAuth: []
      parameters:
        - name: contest_id
          in: path
          required: true
          schema:
            type: integer
        - name: handle
          in: path
          required: true
          schema:
            type: string
      responses:
        204:
          description: Маппинг удалён
        401:
          $ref: '#/components/responses/ErrorResponse'
        404:
          $ref: '#/components/responses/ErrorResponse'

  /api/regatta/contests/{contest_id}:
    get:
      operationId: getRegattaContestStandings
      parameters:
        - name: contest_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          $ref: '#/components/responses/RegattaContestStandings'

  /api/cfcontest/{contest_id}:
    get:
      operationId: getContest
      parameters:
        - name: contest_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          $ref: '#/components/responses/GetContestResponse'

  /api/admin/auth/login:
    post:
      operationId: postAdminAuthLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
      responses:
        200:
          description: Успешная авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminLoginResponse'
        401:
          $ref: '#/components/responses/ErrorResponse'
        503:
          $ref: '#/components/responses/ErrorResponse'

  /api/admin/me:
    get:
      operationId: getAdminMe
      security:
        - bearerAuth: []
      responses:
        200:
          description: Информация о текущем админе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminMeResponse'
        401:
          $ref: '#/components/responses/ErrorResponse'

  /api/admin/timetables/{contest_id}:
    get:
      operationId: getAdminTimetable
      security:
        - bearerAuth: []
      parameters:
        - name: contest_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Расписание туров контеста
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimetableView'
        401:
          $ref: '#/components/responses/ErrorResponse'
        404:
          $ref: '#/components/responses/ErrorResponse'
    put:
      operationId: putAdminTimetable
      security:
        - bearerAuth: []
      parameters:
        - name: contest_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveTimetableRequest'
      responses:
        200:
          description: Сохранённое расписание туров контеста
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimetableView'
        400:
          $ref: '#/components/responses/ErrorResponse'
        401:
          $ref: '#/components/responses/ErrorResponse'
    delete:
      operationId: deleteAdminTimetable
      security:
        - bearerAuth: []
      parameters:
        - name: contest_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        204:
          description: Расписание удалено
        401:
          $ref: '#/components/responses/ErrorResponse'
        404:
          $ref: '#/components/responses/ErrorResponse'

  /api/admin/timetables/{contest_id}/active-tour/duration:
    patch:
      operationId: patchAdminTimetableActiveTourDuration
      security:
        - bearerAuth: []
      parameters:
        - name: contest_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateActiveTourDurationRequest'
      responses:
        200:
          description: Расписание после изменения длительности активного тура
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimetableView'
        400:
          $ref: '#/components/responses/ErrorResponse'
        401:
          $ref: '#/components/responses/ErrorResponse'
        404:
          $ref: '#/components/responses/ErrorResponse'

  /api/admin/timetables/{contest_id}/advance:
    post:
      operationId: postAdminTimetableAdvance
      security:
        - bearerAuth: []
      parameters:
        - name: contest_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Расписание после продвижения (завершение активного сегмента и/или старт следующего слота)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimetableView'
        400:
          $ref: '#/components/responses/ErrorResponse'
        401:
          $ref: '#/components/responses/ErrorResponse'
        404:
          $ref: '#/components/responses/ErrorResponse'
        409:
          $ref: '#/components/responses/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    RegattaContestStandings:
      description: Список студентов с результатами контеста
      content:
        application/json:
          schema:
            type: object
            required:
              - contest_name
              - contest_id
              - current_tour_start_time
              - current_tour_duration
              - rows
              - events
            properties:
              contest_name:
                type: string
                description: Название контеста
              contest_id:
                type: integer
                description: Идентификатор контеста
              current_tour_start_time:
                type: integer
                description: Время начала последнего тура (Unix timestamp в секундах)
              current_tour_duration:
                type: integer
                description: Длительность последнего тура в минутах
              rows:
                type: array
                items:
                  $ref: '#/components/schemas/RegattaContestRow'
              events:
                type: array
                items:
                  $ref: '#/components/schemas/RegattaEvent'

    GetContestResponse:
      description: Информацию о результатах контеста
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Contest'

    ErrorResponse:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    RegattaContestRow:
      type: object
      required:
        - display_name
        - team_number
        - total_score
        - solved_problems
        - problem_results
        - user_id
      properties:
        display_name:
          type: string
          description: Имя и фамилия студента одной строкой
          example: "Иван Иванов"
        team_number:
          type: integer
          description: Номер команды
        total_score:
          type: number
          description: Общий счёт в баллах
          example: 350
        solved_problems:
          type: integer
          description: Количество решенных задач
          example: 7
        problem_results:
          type: array
          description: Результаты по задачам в формате "идентификатор задачи -> баллы", либо <0 за неудачную попытку
          items:
            $ref: '#/components/schemas/ProblemResult'
        user_id:
          type: string
          description: Идентификатор студента
    RegattaEvent:
      type: object
      required:
        - type
        - time_sec
        - participant_id
        - display_name
        - problem_code
        - points
      properties:
        type:
          type: string
          enum:
            - problem_solved
        time_sec:
          type: integer
          description: Время посылки в секундах от старта контеста
        participant_id:
          type: string
        display_name:
          type: string
        problem_code:
          type: string
        team_number:
          type: integer
        points:
          type: integer
          description: Баллы, полученные за это решение
        solved_in_time:
          type: boolean
          description: Бонус за решение в срок тура
        first_in_group:
          type: boolean
          description: Бонус за первое решение в группе
    ProblemResult:
      type: object
      required:
        - problem_code
        - score
      description: Результаты по задаче, c буквой задачи, счётом и временем последней попытки. Если score < 0, то попытка неудачна. Время последней посылки отправляется только для удачных посылок
      properties:
        problem_code:
          type: string
        score:
          type: integer
        last_submission_time:
          description: Время последней посылки в секундах с начала контеста
          type: integer
    Contest:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    AdminLoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    AdminLoginResponse:
      type: object
      required:
        - access_token
        - expires_in
      properties:
        access_token:
          type: string
        expires_in:
          type: integer
          description: Время жизни токена в секундах
    AdminMeResponse:
      type: object
      required:
        - username
        - ok
      properties:
        username:
          type: string
        ok:
          type: boolean
    ScheduleSlot:
      type: object
      required:
        - duration
        - kind
      properties:
        duration:
          type: integer
          minimum: 1
          description: Длительность слота в секундах
        kind:
          type: string
          enum: [tour, pause]
    UpdateActiveTourDurationRequest:
      type: object
      required:
        - duration
      properties:
        duration:
          type: integer
          minimum: 1
          description: Новая длительность активного тура или перерыва в секундах
    SaveTimetableRequest:
      type: object
      required:
        - pending_slots
        - auto_start_enabled
      properties:
        pending_slots:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleSlot'
        auto_start_enabled:
          type: boolean
          description: Автозапуск туров по расписанию для этого контеста
    TimelineSegment:
      type: object
      required:
        - kind
        - duration
        - start_time
        - status
        - editable
      properties:
        sequence:
          type: integer
          nullable: true
        pending_index:
          type: integer
          nullable: true
        kind:
          type: string
          enum: [tour, pause]
        round:
          type: integer
          nullable: true
        duration:
          type: integer
        start_time:
          type: integer
          description: Секунды от старта контеста
        status:
          type: string
          enum: [past, active, next, starting, future]
        editable:
          type: boolean
    RegisteredContestItem:
      type: object
      required:
        - contest_id
        - name
        - system
      properties:
        contest_id:
          type: integer
        name:
          type: string
        system:
          type: string
          example: codeforces
    CreateContestRequest:
      type: object
      required:
        - contest_id
      properties:
        contest_id:
          type: integer
        name:
          type: string
          description: Необязательное отображаемое имя; по умолчанию берётся с Codeforces
    CodeforcesHandleItem:
      type: object
      required:
        - handle
        - name
      properties:
        handle:
          type: string
        name:
          type: string
    PutCodeforcesHandlesRequest:
      type: object
      required:
        - handles
      properties:
        handles:
          type: array
          items:
            $ref: '#/components/schemas/CodeforcesHandleItem'
    TimetableView:
      type: object
      required:
        - contest_id
        - server_now
        - elapsed_seconds
        - timeline_segments
        - pending_slots
        - auto_start_enabled
        - auto_start_available
      properties:
        contest_id:
          type: integer
        contest_start_time:
          type: string
          format: date-time
        server_now:
          type: string
          format: date-time
        elapsed_seconds:
          type: integer
        timeline_segments:
          type: array
          items:
            $ref: '#/components/schemas/TimelineSegment'
        pending_slots:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleSlot'
        next_tour_number:
          type: integer
          nullable: true
        auto_start_enabled:
          type: boolean
        auto_start_available:
          type: boolean
          description: На сервере включён фоновый timetable_sync